General cleanup on checkout.c, added getline function in linux-functions.h
For all the functions that are generally avalliable on Linux through libgcc, place them into the linux-functions.h file, so I may use them on Windows as well. Additionally, have cleaned up some errors in checkout.c and improved handling of duplicate instrument numbers. What needs to still be done is improve the overwriting of files and generalizing the task.